Add async call test cases
