+++ tests/jballerina-integration-test/src/test/java/org/ballerinalang/test/observability/tracing/OOTBTracingTestCase.java
@@ -630,107 +630,4 @@ public class OOTBTracingTestCase extends TracingBaseTest {
             ));
         });
     }
-
-    @Test
-    public void testAsyncCall() throws Exception {
-        HttpResponse httpResponse = HttpClientRequest.doGet("http://localhost:9097/test-service/resource-1");
-        Assert.assertEquals(httpResponse.getResponseCode(), 200);
-        Assert.assertEquals(httpResponse.getData(), "Hello, World! from resource one");
-        Thread.sleep(1000);
-
-        final String span1Position = "07_ootb_async.bal:27:5";
-        final String span2Position = "07_ootb_async.bal:31:35";
-        final String span3Position = "07_ootb_async.bal:44:35";
-        final String span4Position = "07_ootb_async.bal:56:20";
-
-        List<BMockSpan> spans = this.getFinishedSpans("testServiceSeven");
-        Assert.assertEquals(spans.size(), 4);
-        Assert.assertEquals(spans.stream()
-                        .map(span -> span.getTags().get("src.position"))
-                        .collect(Collectors.toSet()),
-                new HashSet<>(Arrays.asList(span1Position, span2Position, span3Position, span4Position)));
-        Assert.assertEquals(spans.stream().filter(bMockSpan -> bMockSpan.getParentId() == 0).count(), 1);
-
-        Optional<BMockSpan> span1 = spans.stream()
-                .filter(bMockSpan -> Objects.equals(bMockSpan.getTags().get("src.position"), span1Position))
-                .findFirst();
-        Assert.assertTrue(span1.isPresent());
-        long traceId = span1.get().getTraceId();
-        span1.ifPresent(span -> {
-            Assert.assertTrue(spans.stream().noneMatch(mockSpan -> mockSpan.getTraceId() == traceId
-                    && mockSpan.getSpanId() == span.getParentId()));
-            Assert.assertEquals(span.getOperationName(), "resourceOne");
-            Assert.assertEquals(span.getTags(), toMap(
-                    new AbstractMap.SimpleEntry<>("span.kind", "server"),
-                    new AbstractMap.SimpleEntry<>("src.module", "ballerina-test/tracingservices:0.0.1"),
-                    new AbstractMap.SimpleEntry<>("src.position", span1Position),
-                    new AbstractMap.SimpleEntry<>("src.entry_point.resource", "true"),
-                    new AbstractMap.SimpleEntry<>("http.url", "/test-service/resource-1"),
-                    new AbstractMap.SimpleEntry<>("http.method", "GET"),
-                    new AbstractMap.SimpleEntry<>("protocol", "http"),
-                    new AbstractMap.SimpleEntry<>("service", "testServiceSeven"),
-                    new AbstractMap.SimpleEntry<>("resource", "resourceOne"),
-                    new AbstractMap.SimpleEntry<>("connector_name", "http")
-            ));
-        });
-
-        Optional<BMockSpan> span2 = spans.stream()
-                .filter(bMockSpan -> Objects.equals(bMockSpan.getTags().get("src.position"), span2Position))
-                .findFirst();
-        Assert.assertTrue(span2.isPresent());
-        span2.ifPresent(span -> {
-            Assert.assertEquals(span.getTraceId(), traceId);
-            Assert.assertEquals(span.getParentId(), span1.get().getSpanId());
-            Assert.assertEquals(span.getOperationName(), "calculateSumWithObservableFunction");
-            Assert.assertEquals(span.getTags(), toMap(
-                    new AbstractMap.SimpleEntry<>("span.kind", "client"),
-                    new AbstractMap.SimpleEntry<>("src.module", "ballerina-test/tracingservices:0.0.1"),
-                    new AbstractMap.SimpleEntry<>("src.position", span2Position),
-                    new AbstractMap.SimpleEntry<>("service", "testServiceSeven"),
-                    new AbstractMap.SimpleEntry<>("resource", "resourceOne"),
-                    new AbstractMap.SimpleEntry<>("function", "calculateSumWithObservableFunction")
-            ));
-        });
-
-        Optional<BMockSpan> span3 = spans.stream()
-                .filter(bMockSpan -> Objects.equals(bMockSpan.getTags().get("src.position"), span3Position))
-                .findFirst();
-        Assert.assertTrue(span3.isPresent());
-        span3.ifPresent(span -> {
-            Assert.assertEquals(span.getTraceId(), traceId);
-            Assert.assertEquals(span.getParentId(), span1.get().getSpanId());
-            Assert.assertEquals(span.getOperationName(), "ballerina-test/tracingservices/ObservableAdder:getSum");
-            Assert.assertEquals(span.getTags(), toMap(
-                    new AbstractMap.SimpleEntry<>("span.kind", "client"),
-                    new AbstractMap.SimpleEntry<>("src.module", "ballerina-test/tracingservices:0.0.1"),
-                    new AbstractMap.SimpleEntry<>("src.position", span3Position),
-                    new AbstractMap.SimpleEntry<>("service", "testServiceSeven"),
-                    new AbstractMap.SimpleEntry<>("resource", "resourceOne"),
-                    new AbstractMap.SimpleEntry<>("function", "getSum"),
-                    new AbstractMap.SimpleEntry<>("object_name", "ballerina-test/tracingservices/ObservableAdder")
-            ));
-        });
-
-        Optional<BMockSpan> span4 = spans.stream()
-                .filter(bMockSpan -> Objects.equals(bMockSpan.getTags().get("src.position"), span4Position))
-                .findFirst();
-        Assert.assertTrue(span4.isPresent());
-        span4.ifPresent(span -> {
-            Assert.assertEquals(span.getTraceId(), traceId);
-            Assert.assertEquals(span.getParentId(), span1.get().getSpanId());
-            Assert.assertEquals(span.getOperationName(), "ballerina/http/Caller:respond");
-            Assert.assertEquals(span.getTags(), toMap(
-                    new AbstractMap.SimpleEntry<>("span.kind", "client"),
-                    new AbstractMap.SimpleEntry<>("src.module", "ballerina-test/tracingservices:0.0.1"),
-                    new AbstractMap.SimpleEntry<>("src.position", span4Position),
-                    new AbstractMap.SimpleEntry<>("src.remote", "true"),
-                    new AbstractMap.SimpleEntry<>("http.status_code", "200"),
-                    new AbstractMap.SimpleEntry<>("http.status_code_group", "2xx"),
-                    new AbstractMap.SimpleEntry<>("service", "testServiceSeven"),
-                    new AbstractMap.SimpleEntry<>("resource", "resourceOne"),
-                    new AbstractMap.SimpleEntry<>("connector_name", "ballerina/http/Caller"),
-                    new AbstractMap.SimpleEntry<>("action", "respond")
-            ));
-        });
-    }
 }
